requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
pymupdf==1.24.5
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84
pyzbar==0.1.9
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.134.0
python-dotenv==1.0.1
requests==2.32.3
